refactor: relocate provider, axios-sdk and login function(into axios-sdk) and update its imports
