+++ app/layout.tsx
@@ -1,4 +1,4 @@
-import Providers from '@/utils/provider'
+import Providers from '@/components/provider'
 import './globals.css'
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
