--- utils/axios-sdk.ts
@@ -1,5 +1,0 @@
-import axios from "axios";
-
-export const api = axios.create({
-    baseURL: process.env.NEXT_PUBLIC_REST_API_URL,
-})
